Tests/LibWeb/Text/expected/css/flex-basis-comparison-function.txt

image: 29, 763.859375
info: 888.859375, 50, 457.140625
sticky: 903.859375, 112, 427.140625
